--- packages/common/src/projects/HubProjects.ts
@@ -1,339 +1,0 @@
-import { IUserRequestOptions, UserSession } from "@esri/arcgis-rest-auth";
-
-// Note - we separate these imports so we can cleanly spy on things in tests
-import {
-  createModel,
-  fetchModelFromItem,
-  getModel,
-  updateModel,
-} from "../models";
-import {
-  constructSlug,
-  getItemBySlug,
-  getUniqueSlug,
-  setSlugKeyword,
-} from "../items/slugs";
-import {
-  IModel,
-  isGuid,
-  cloneObject,
-  getItemThumbnailUrl,
-  unique,
-  mapBy,
-  getProp,
-  getFamily,
-  IHubRequestOptions,
-  getItemHomeUrl,
-  interpolate,
-  EditorConfigType,
-  UiSchemaElementOptions,
-  IEditorConfig,
-  deepFind,
-} from "../index";
-import {
-  IItem,
-  IUserItemOptions,
-  removeItem,
-  getItem,
-} from "@esri/arcgis-rest-portal";
-import { IRequestOptions } from "@esri/arcgis-rest-request";
-
-import { IPropertyMap, PropertyMapper } from "../core/_internal/PropertyMapper";
-import { IHubProject } from "../core/types";
-import { IHubSearchResult } from "../search";
-import { parseInclude } from "../search/_internal/parseInclude";
-import { fetchItemEnrichments } from "../items/_enrichments";
-import { getHubRelativeUrl } from "../content/_internal";
-import { DEFAULT_PROJECT, DEFAULT_PROJECT_MODEL } from "./defaults";
-import { getBasePropertyMap } from "../core/_internal/getBasePropertyMap";
-import {
-  HubProjectEditUiSchema,
-  HubProjectCreateUiSchema,
-  HubProjectSchema,
-} from "../core/schemas";
-import { filterSchemaToUiSchema } from "../core/schemas/internal";
-import { applyUiSchemaElementOptions } from "../core/schemas/internal/applyUiSchemaElementOptions";
-
-/**
- * Returns an Array of IPropertyMap objects
- * that define the projection of properties from a IModel to an IHubProject
- * @returns
- */
-function getPropertyMap(): IPropertyMap[] {
-  const map = getBasePropertyMap();
-
-  // Type specific mappings
-  map.push({ objectKey: "status", modelKey: "data.status" });
-  map.push({ objectKey: "catalog", modelKey: "data.catalog" });
-  map.push({ objectKey: "permissions", modelKey: "data.permissions" });
-  map.push({ objectKey: "contacts", modelKey: "data.contacts" });
-  map.push({ objectKey: "timeline", modelKey: "data.timeline" });
-
-  return map;
-}
-
-/**
- * @private
- * Create a new Hub Project item
- *
- * Minimal properties are name and org
- *
- * @param project
- * @param requestOptions
- */
-export async function createProject(
-  partialProject: Partial<IHubProject>,
-  requestOptions: IUserRequestOptions
-): Promise<IHubProject> {
-  // merge incoming with the default
-  // this expansion solves the typing somehow
-  const project = { ...DEFAULT_PROJECT, ...partialProject };
-
-  // Create a slug from the title if one is not passed in
-  if (!project.slug) {
-    project.slug = constructSlug(project.name, project.orgUrlKey);
-  }
-  // Ensure slug is  unique
-  project.slug = await getUniqueSlug({ slug: project.slug }, requestOptions);
-  // add slug to keywords
-  project.typeKeywords = setSlugKeyword(project.typeKeywords, project.slug);
-  // Map project object onto a default project Model
-  const mapper = new PropertyMapper<Partial<IHubProject>>(getPropertyMap());
-  // create model from object, using the default model as a starting point
-  let model = mapper.objectToModel(project, cloneObject(DEFAULT_PROJECT_MODEL));
-  // create the item
-  model = await createModel(model, requestOptions);
-  // map the model back into a IHubProject
-  let newProject = mapper.modelToObject(model, {});
-  newProject = computeProps(model, newProject, requestOptions);
-  // and return it
-  return newProject as IHubProject;
-}
-
-/**
- * @private
- * Update a Hub Project
- * @param project
- * @param requestOptions
- */
-export async function updateProject(
-  project: IHubProject,
-  requestOptions: IUserRequestOptions
-): Promise<IHubProject> {
-  // verify that the slug is unique, excluding the current project
-  project.slug = await getUniqueSlug(
-    { slug: project.slug, existingId: project.id },
-    requestOptions
-  );
-  // get the backing item & data
-  const model = await getModel(project.id, requestOptions);
-  // create the PropertyMapper
-  const mapper = new PropertyMapper<Partial<IHubProject>>(getPropertyMap());
-  // Note: Although we are fetching the model, and applying changes onto it,
-  // we are not attempting to handle "concurrent edit" conflict resolution
-  // but this is where we would apply that sort of logic
-  const modelToUpdate = mapper.objectToModel(project, model);
-  // update the backing item
-  const updatedModel = await updateModel(modelToUpdate, requestOptions);
-  // now map back into a project and return that
-  let updatedProject = mapper.modelToObject(updatedModel, project);
-  updatedProject = computeProps(model, updatedProject, requestOptions);
-  // the casting is needed because modelToObject returns a `Partial<T>`
-  // where as this function returns a `T`
-  return updatedProject as IHubProject;
-}
-
-/**
- * @private
- * Get a Hub Project by id or slug
- * @param identifier item id or slug
- * @param requestOptions
- */
-export function fetchProject(
-  identifier: string,
-  requestOptions: IRequestOptions
-): Promise<IHubProject> {
-  let getPrms;
-  if (isGuid(identifier)) {
-    // get item by id
-    getPrms = getItem(identifier, requestOptions);
-  } else {
-    getPrms = getItemBySlug(identifier, requestOptions);
-  }
-  return getPrms.then((item) => {
-    if (!item) return null;
-    return convertItemToProject(item, requestOptions);
-  });
-}
-
-/**
- * @private
- * Remove a Hub Project
- * @param id
- * @param requestOptions
- */
-export async function deleteProject(
-  id: string,
-  requestOptions: IUserRequestOptions
-): Promise<void> {
-  const ro = { ...requestOptions, ...{ id } } as IUserItemOptions;
-  await removeItem(ro);
-  return;
-}
-
-/**
- * @private
- * Convert an Hub Project Item into a Hub Project, fetching any additional
- * information that may be required
- * @param item
- * @param auth
- * @returns
- */
-export async function convertItemToProject(
-  item: IItem,
-  requestOptions: IRequestOptions
-): Promise<IHubProject> {
-  const model = await fetchModelFromItem(item, requestOptions);
-  // TODO: In the future we will handle the boundary fetching from resource
-  const mapper = new PropertyMapper<Partial<IHubProject>>(getPropertyMap());
-  const prj = mapper.modelToObject(model, {}) as IHubProject;
-  return computeProps(model, prj, requestOptions);
-}
-
-/**
- * Given a model and a project, set various computed properties that can't be directly mapped
- * @param model
- * @param project
- * @param requestOptions
- * @returns
- */
-function computeProps(
-  model: IModel,
-  project: Partial<IHubProject>,
-  requestOptions: IRequestOptions
-): IHubProject {
-  let token: string;
-  if (requestOptions.authentication) {
-    const session: UserSession = requestOptions.authentication as UserSession;
-    token = session.token;
-  }
-  // thumbnail url
-  project.thumbnailUrl = getItemThumbnailUrl(model.item, requestOptions, token);
-  // Handle Dates
-  project.createdDate = new Date(model.item.created);
-  project.createdDateSource = "item.created";
-  project.updatedDate = new Date(model.item.modified);
-  project.updatedDateSource = "item.modified";
-  // cast b/c this takes a partial but returns a full project
-  return project as IHubProject;
-}
-
-/**
- * @private
- * Fetch project specific enrichments
- * @param item
- * @param include
- * @param requestOptions
- * @returns
- */
-export async function enrichProjectSearchResult(
-  item: IItem,
-  include: string[],
-  requestOptions: IHubRequestOptions
-): Promise<IHubSearchResult> {
-  // Create the basic structure
-  const result: IHubSearchResult = {
-    access: item.access,
-    id: item.id,
-    type: item.type,
-    name: item.title,
-    owner: item.owner,
-    typeKeywords: item.typeKeywords,
-    tags: item.tags,
-    categories: item.categories,
-    summary: item.snippet || item.description,
-    createdDate: new Date(item.created),
-    createdDateSource: "item.created",
-    updatedDate: new Date(item.modified),
-    updatedDateSource: "item.modified",
-    family: getFamily(item.type),
-    links: {
-      self: "not-implemented",
-      siteRelative: "not-implemented",
-      thumbnail: "not-implemented",
-    },
-  };
-
-  // default includes
-  const DEFAULTS: string[] = [];
-  // merge includes
-  include = [...DEFAULTS, ...include].filter(unique);
-  // Parse the includes into a valid set of enrichments
-  const specs = include.map(parseInclude);
-  // Extract out the low-level enrichments needed
-  const enrichments = mapBy("enrichment", specs).filter(unique);
-  // fetch the enrichments
-  let enriched = {};
-  if (enrichments.length) {
-    // TODO: Look into caching for the requests in fetchItemEnrichments
-    enriched = await fetchItemEnrichments(item, enrichments, requestOptions);
-  }
-  // map the enriched props onto the result
-  specs.forEach((spec) => {
-    result[spec.prop] = getProp(enriched, spec.path);
-  });
-
-  // Handle links
-  // TODO: Link handling should be an enrichment
-  result.links.thumbnail = getItemThumbnailUrl(item, requestOptions);
-  result.links.self = getItemHomeUrl(result.id, requestOptions);
-  result.links.siteRelative = getHubRelativeUrl(
-    result.type,
-    result.id,
-    item.typeKeywords
-  );
-
-  return result;
-}
-
-/**
- * Get the editor config for for the HubProject entity.
- * @param i18nScope Translation scope to be interpolated into the schemas
- * @param type
- * @param options Optional hash of Element component options
- * @returns
- */
-export async function getHubProjectEditorConfig(
-  i18nScope: string,
-  type: EditorConfigType,
-  options: UiSchemaElementOptions[] = []
-): Promise<IEditorConfig> {
-  // schema is always the entire schema
-  let schema = cloneObject(HubProjectSchema);
-  // uiSchema is the complete (edit) schema, unless otherwise specified
-  let uiSchema = cloneObject(HubProjectEditUiSchema);
-  // by default we don't need to filter the schema b/c it's the entire schema
-  let filterSchema = false;
-
-  // if another schema is requested, we need to use that UI schema
-  // and the subset the overall schema down to just the properties
-  // used in the UI schema
-  switch (type) {
-    case "create":
-      uiSchema = cloneObject(HubProjectCreateUiSchema);
-      filterSchema = true;
-      break;
-  }
-
-  if (filterSchema) {
-    // filter out properties not used in the UI schema
-    schema = filterSchemaToUiSchema(schema, uiSchema);
-  }
-
-  // interpolate the i18n scope into the uiSchema
-  uiSchema = interpolate(uiSchema, { i18nScope });
-  // apply the options
-  uiSchema = applyUiSchemaElementOptions(uiSchema, options);
-
-  return Promise.resolve({ schema, uiSchema });
-}
