+++ packages/common/src/projects/index.ts
@@ -1,2 +1,3 @@
-export * from "./HubProjects";
 export * from "./HubProject";
+export * from "./edit";
+export * from "./fetch";
